fix: resolve 7 mypy strict errors in executor.py, cli.py, models.py
- models.py: change poll_interval_seconds type int|None → float|None
- executor.py #553: add type:ignore[no-any-return] for pd.concat
- executor.py #693-696: fix poll_interval type narrowing with float default
- executor.py #704,719,722: add assert self.state is not None before loop body

+++ ztb/execution/models.py
@@ -84,7 +84,7 @@ class ExecRunConfig:
     dry_run: bool = False
     once: bool = False
     loop: bool | None = None
-    poll_interval_seconds: int | None = None
+    poll_interval_seconds: float | None = None
     initial_cash: float = 100_000.0
     commission: float = 0.0005
     slippage: float = 0.0005
